refactor: move the create_images class to main
