+++ src/utils/assets/create_images.py
@@ -1,7 +1,7 @@
 from pygame import image, transform
 from typing import Text, Tuple
 
-def create_image(path: Text, size: Tuple[int, int]):
+def create_image(path: Text, size: Tuple[float, float]):
     return transform.scale(
         image.load(path), size
     )
